Move code from execute_block to runtime_externals
