+++ src/verify/execute_block.rs
@@ -39,11 +39,16 @@
 //! block in order to continue.
 //!
 
-use crate::{executor, header, trie::calculate_root, util};
+use crate::{
+    executor::{self, runtime_externals},
+    header,
+    trie::calculate_root,
+    util,
+};
 
 use alloc::{string::String, vec::Vec};
-use core::{iter, slice};
-use hashbrown::{HashMap, HashSet};
+use core::iter;
+use hashbrown::HashMap;
 
 /// Configuration for an unsealed block verification.
 pub struct Config<'a, TBody> {
@@ -85,26 +90,20 @@ pub struct Success {
 #[derive(Debug, Clone, derive_more::Display)]
 pub enum Error {
     /// Error while executing the Wasm virtual machine.
-    #[display(fmt = "Error while executing Wasm VM: {}", error)]
-    WasmVm {
-        /// Error that happened.
-        error: executor::Error,
-        /// Concatenation of all the log messages printed by the runtime.
-        logs: String,
-    },
+    #[display(fmt = "{}", _0)]
+    WasmVm(runtime_externals::Error),
     /// Output of `Core_execute_block` wasn't empty.
     NonEmptyOutput,
-    /// Size of the logs generated by the runtime exceeds the limit.
-    LogsTooLong,
 }
 
 /// Verifies whether a block is valid.
 pub fn execute_block<'a>(
     config: Config<'a, impl ExactSizeIterator<Item = impl AsRef<[u8]> + Clone> + Clone>,
 ) -> Verify {
-    let vm = config
-        .parent_runtime
-        .run_vectored("Core_execute_block", {
+    let vm = runtime_externals::run_vectored(
+        config.parent_runtime,
+        "Core_execute_block",
+        {
             // The `Code_execute_block` function expects a SCALE-encoded `(header, body)`
             // where `body` is a `Vec<Vec<u8>>`. We perform the encoding manually to avoid
             // performing redundant data copies.
@@ -120,21 +119,14 @@ pub fn execute_block<'a>(
                 .map(|b| either::Right(either::Left(b)))
                 .chain(iter::once(either::Right(either::Right(encoded_body_len))))
                 .chain(body.map(either::Left))
-        })
-        .unwrap()
-        .into();
+        },
+        config.top_trie_root_calculation_cache,
+    )
+    .unwrap();
 
-    VerifyInner {
-        vm,
-        top_trie_changes: Default::default(),
-        offchain_storage_changes: Default::default(),
-        top_trie_root_calculation_cache: Some(
-            config.top_trie_root_calculation_cache.unwrap_or_default(),
-        ),
-        root_calculation: None,
-        logs: String::new(),
-    }
-    .run()
+    // TODO: shouldn't unwrap ^
+
+    Verify::from_inner(vm)
 }
 
 /// Current state of the verification.
@@ -150,214 +142,86 @@ pub enum Verify {
     NextKey(NextKey),
 }
 
+impl Verify {
+    fn from_inner(inner: runtime_externals::RuntimeExternalsVm) -> Self {
+        match inner {
+            runtime_externals::RuntimeExternalsVm::Finished(Ok(success)) => {
+                if !success.virtual_machine.value().is_empty() {
+                    return Verify::Finished(Err(Error::NonEmptyOutput));
+                }
+
+                return Verify::Finished(Ok(Success {
+                    parent_runtime: success.virtual_machine.into_prototype(),
+                    storage_top_trie_changes: success.storage_top_trie_changes,
+                    offchain_storage_changes: success.offchain_storage_changes,
+                    top_trie_root_calculation_cache: success.top_trie_root_calculation_cache,
+                    logs: success.logs,
+                }));
+            }
+            runtime_externals::RuntimeExternalsVm::Finished(Err(err)) => {
+                Verify::Finished(Err(Error::WasmVm(err)))
+            }
+            runtime_externals::RuntimeExternalsVm::StorageGet(inner) => {
+                Verify::StorageGet(StorageGet(inner))
+            }
+            runtime_externals::RuntimeExternalsVm::PrefixKeys(inner) => {
+                Verify::PrefixKeys(PrefixKeys(inner))
+            }
+            runtime_externals::RuntimeExternalsVm::NextKey(inner) => {
+                Verify::NextKey(NextKey(inner))
+            }
+        }
+    }
+}
+
 /// Loading a storage value is required in order to continue.
 #[must_use]
-pub struct StorageGet {
-    inner: VerifyInner,
-}
+pub struct StorageGet(runtime_externals::StorageGet);
 
 impl StorageGet {
     /// Returns the key whose value must be passed to [`StorageGet::inject_value`].
     pub fn key<'a>(&'a self) -> impl Iterator<Item = impl AsRef<[u8]> + 'a> + 'a {
-        match &self.inner.vm {
-            executor::WasmVm::ExternalStorageGet(req) => {
-                either::Either::Left(iter::once(either::Either::Left(req.key())))
-            }
-            executor::WasmVm::ExternalStorageAppend(req) => {
-                either::Either::Left(iter::once(either::Either::Left(req.key())))
-            }
-
-            executor::WasmVm::ExternalStorageRoot(_) => {
-                if let calculate_root::RootMerkleValueCalculation::StorageValue(value_request) =
-                    self.inner.root_calculation.as_ref().unwrap()
-                {
-                    struct One(u8);
-                    impl AsRef<[u8]> for One {
-                        fn as_ref(&self) -> &[u8] {
-                            slice::from_ref(&self.0)
-                        }
-                    }
-                    either::Either::Right(value_request.key().map(One).map(either::Either::Right))
-                } else {
-                    // We only create a `StorageGet` if the state is `StorageValue`.
-                    panic!()
-                }
-            }
-
-            executor::WasmVm::ExternalStorageChangesRoot(_) => {
-                either::Either::Left(iter::once(either::Either::Left(&b":changes_trie"[..])))
-            }
-
-            // We only create a `StorageGet` if the state is one of the above.
-            _ => unreachable!(),
-        }
+        self.0.key()
     }
 
     /// Returns the key whose value must be passed to [`StorageGet::inject_value`].
     ///
     /// This method is a shortcut for calling `key` and concatenating the returned slices.
     pub fn key_as_vec(&self) -> Vec<u8> {
-        self.key().fold(Vec::new(), |mut a, b| {
-            a.extend_from_slice(b.as_ref());
-            a
-        })
+        self.0.key_as_vec()
     }
 
     /// Injects the corresponding storage value.
     // TODO: `value` parameter should be something like `Iterator<Item = impl AsRef<[u8]>`
-    pub fn inject_value(mut self, value: Option<&[u8]>) -> Verify {
-        match self.inner.vm {
-            executor::WasmVm::ExternalStorageGet(req) => {
-                // TODO: should actually report the offset and max_size in the API
-                self.inner.vm = req.resume_full_value(value);
-            }
-            executor::WasmVm::ExternalStorageAppend(req) => {
-                let mut value = value.map(|v| v.to_vec()).unwrap_or_default();
-                // TODO: could be less overhead?
-                append_to_storage_value(&mut value, req.value());
-                self.inner
-                    .top_trie_changes
-                    .insert(req.key().to_vec(), Some(value.clone()));
-                self.inner.vm = req.resume();
-            }
-            executor::WasmVm::ExternalStorageRoot(_) => {
-                if let calculate_root::RootMerkleValueCalculation::StorageValue(value_request) =
-                    self.inner.root_calculation.take().unwrap()
-                {
-                    self.inner.root_calculation = Some(value_request.inject(value));
-                } else {
-                    // We only create a `StorageGet` if the state is `StorageValue`.
-                    panic!()
-                }
-            }
-            executor::WasmVm::ExternalStorageChangesRoot(req) => {
-                if value.is_none() {
-                    self.inner.vm = req.resume(None);
-                } else {
-                    // TODO: this is probably one of the most complicated things to implement
-                    todo!()
-                }
-            }
-
-            // We only create a `StorageGet` if the state is one of the above.
-            _ => unreachable!(),
-        };
-
-        self.inner.run()
+    pub fn inject_value(self, value: Option<&[u8]>) -> Verify {
+        Verify::from_inner(self.0.inject_value(value))
     }
 }
 
 /// Fetching the list of keys with a given prefix is required in order to continue.
 #[must_use]
-pub struct PrefixKeys {
-    inner: VerifyInner,
-}
+pub struct PrefixKeys(runtime_externals::PrefixKeys);
 
 impl PrefixKeys {
     /// Returns the prefix whose keys to load.
     pub fn prefix(&self) -> &[u8] {
-        match &self.inner.vm {
-            executor::WasmVm::ExternalStorageClearPrefix(req) => req.prefix(),
-            executor::WasmVm::ExternalStorageRoot { .. } => &[],
-
-            // We only create a `PrefixKeys` if the state is one of the above.
-            _ => unreachable!(),
-        }
+        self.0.prefix()
     }
 
     /// Injects the list of keys.
-    pub fn inject_keys(mut self, keys: impl Iterator<Item = impl AsRef<[u8]>>) -> Verify {
-        match self.inner.vm {
-            executor::WasmVm::ExternalStorageClearPrefix(req) => {
-                // TODO: use prefix_remove_update once optimized
-                //top_trie_root_calculation_cache.prefix_remove_update(storage_key);
-
-                for key in keys {
-                    self.inner
-                        .top_trie_root_calculation_cache
-                        .as_mut()
-                        .unwrap()
-                        .storage_value_update(key.as_ref(), false);
-                    self.inner
-                        .top_trie_changes
-                        .insert(key.as_ref().to_vec(), None);
-                }
-                // TODO: O(n) complexity here
-                for (key, value) in self.inner.top_trie_changes.iter_mut() {
-                    if !key.starts_with(req.prefix()) {
-                        continue;
-                    }
-                    if value.is_none() {
-                        continue;
-                    }
-                    self.inner
-                        .top_trie_root_calculation_cache
-                        .as_mut()
-                        .unwrap()
-                        .storage_value_update(key, false);
-                    *value = None;
-                }
-                self.inner.vm = req.resume();
-            }
-
-            executor::WasmVm::ExternalStorageRoot { .. } => {
-                if let calculate_root::RootMerkleValueCalculation::AllKeys(all_keys) =
-                    self.inner.root_calculation.take().unwrap()
-                {
-                    // TODO: overhead
-                    let mut list = keys
-                        .filter(|v| {
-                            self.inner
-                                .top_trie_changes
-                                .get(v.as_ref())
-                                .map_or(true, |v| v.is_some())
-                        })
-                        .map(|v| v.as_ref().to_vec())
-                        .collect::<HashSet<_, fnv::FnvBuildHasher>>();
-                    // TODO: slow to iterate over everything?
-                    for (key, value) in self.inner.top_trie_changes.iter() {
-                        if value.is_none() {
-                            continue;
-                        }
-                        list.insert(key.clone());
-                    }
-                    self.inner.root_calculation =
-                        Some(all_keys.inject(list.into_iter().map(|k| k.into_iter())));
-                } else {
-                    // We only create a `PrefixKeys` if the state is `AllKeys`.
-                    panic!()
-                }
-            }
-
-            // We only create a `PrefixKeys` if the state is one of the above.
-            _ => unreachable!(),
-        };
-
-        self.inner.run()
+    pub fn inject_keys(self, keys: impl Iterator<Item = impl AsRef<[u8]>>) -> Verify {
+        Verify::from_inner(self.0.inject_keys(keys))
     }
 }
 
 /// Fetching the key that follows a given one is required in order to continue.
 #[must_use]
-pub struct NextKey {
-    inner: VerifyInner,
-
-    /// If `Some`, ask for the key inside of this field rather than the one of `inner`. Used in
-    /// corner-case situations where the key provided by the user has been erased from storage.
-    key_overwrite: Option<Vec<u8>>,
-}
+pub struct NextKey(runtime_externals::NextKey);
 
 impl NextKey {
     /// Returns the key whose next key must be passed back.
     pub fn key(&self) -> &[u8] {
-        if let Some(key_overwrite) = &self.key_overwrite {
-            return key_overwrite;
-        }
-
-        match &self.inner.vm {
-            executor::WasmVm::ExternalStorageNextKey(req) => req.key(),
-            _ => unreachable!(),
-        }
+        self.0.key()
     }
 
     /// Injects the key.
@@ -366,329 +230,7 @@ impl NextKey {
     ///
     /// Panics if the key passed as parameter isn't strictly superior to the requested key.
     ///
-    pub fn inject_key(mut self, key: Option<impl AsRef<[u8]>>) -> Verify {
-        let key = key.as_ref().map(|k| k.as_ref());
-
-        match self.inner.vm {
-            executor::WasmVm::ExternalStorageNextKey(req) => {
-                let requested_key = if let Some(key_overwrite) = &self.key_overwrite {
-                    &key_overwrite[..]
-                } else {
-                    req.key()
-                };
-
-                if let Some(key) = key {
-                    assert!(key > requested_key);
-                }
-
-                // The next key can be either the one passed by the user or one key in the current
-                // pending storage changes that has been inserted during the verification.
-                // As such, find the "next key" in the list of overlay changes.
-                // TODO: not optimized in terms of searching time ; should really be a BTreeMap or something
-                let in_overlay = self
-                    .inner
-                    .top_trie_changes
-                    .iter()
-                    .map(|(k, v)| (k, v.is_some()))
-                    .filter(|(k, _)| &***k > requested_key)
-                    .min_by_key(|(k, _)| *k);
-
-                let outcome = match (key, in_overlay) {
-                    (Some(a), Some((b, true))) if a <= &b[..] => Some(a),
-                    (Some(a), Some((b, false))) if a < &b[..] => Some(a),
-                    (Some(a), Some((b, false))) => {
-                        debug_assert!(a >= &b[..]);
-                        debug_assert_ne!(&b[..], requested_key);
-
-                        // The next key according to the parent storage has been erased earlier in
-                        // the block execution. It is necessary to ask the user again, this time
-                        // for the key after the one that has been erased.
-                        // This `clone()` is necessary, as `b` borrows from
-                        // `self.inner.top_trie_changes`.
-                        let key_overwrite = Some(b.clone());
-                        self.inner.vm = executor::WasmVm::ExternalStorageNextKey(req);
-                        return Verify::NextKey(NextKey {
-                            inner: self.inner,
-                            key_overwrite,
-                        });
-                    }
-                    (Some(a), Some((b, true))) => {
-                        debug_assert!(a >= &b[..]);
-                        Some(&b[..])
-                    }
-
-                    (Some(a), None) => Some(a),
-                    (None, Some((b, _))) => Some(&b[..]),
-                    (None, None) => None,
-                };
-
-                self.inner.vm = req.resume(outcome.as_ref().map(|v| &v[..]));
-            }
-
-            // We only create a `NextKey` if the state is one of the above.
-            _ => unreachable!(),
-        };
-
-        self.inner.run()
+    pub fn inject_key(self, key: Option<impl AsRef<[u8]>>) -> Verify {
+        Verify::from_inner(self.0.inject_key(key))
     }
-}
-
-/// Implementation detail of the verification. Shared by all the variants of [`Verify`] other
-/// than [`Verify::Finished`].
-struct VerifyInner {
-    /// Virtual machine running the call.
-    vm: executor::WasmVm,
-
-    /// Pending changes to the top storage trie that this block performs.
-    top_trie_changes: HashMap<Vec<u8>, Option<Vec<u8>>, fnv::FnvBuildHasher>,
-
-    /// Pending changes to the offchain storage that this block performs.
-    offchain_storage_changes: HashMap<Vec<u8>, Option<Vec<u8>>, fnv::FnvBuildHasher>,
-
-    /// Cache passed by the user in the [`Config`]. Always `Some` except when we are currently
-    /// calculating the trie state root.
-    top_trie_root_calculation_cache: Option<calculate_root::CalculationCache>,
-
-    /// Trie root calculation in progress.
-    root_calculation: Option<calculate_root::RootMerkleValueCalculation>,
-
-    /// Concatenation of all the log messages generated by the runtime.
-    logs: String,
-}
-
-impl VerifyInner {
-    /// Continues the verification.
-    fn run(mut self) -> Verify {
-        loop {
-            match self.vm {
-                executor::WasmVm::ReadyToRun(r) => self.vm = r.run(),
-
-                executor::WasmVm::Error { error, .. } => {
-                    return Verify::Finished(Err(Error::WasmVm {
-                        error,
-                        logs: self.logs,
-                    }));
-                }
-
-                executor::WasmVm::Finished(finished) => {
-                    if !finished.value().is_empty() {
-                        return Verify::Finished(Err(Error::NonEmptyOutput));
-                    }
-
-                    return Verify::Finished(Ok(Success {
-                        parent_runtime: finished.into_prototype(),
-                        storage_top_trie_changes: self.top_trie_changes,
-                        offchain_storage_changes: self.offchain_storage_changes,
-                        top_trie_root_calculation_cache: self
-                            .top_trie_root_calculation_cache
-                            .unwrap(),
-                        logs: self.logs,
-                    }));
-                }
-
-                executor::WasmVm::ExternalStorageGet(req) => {
-                    if let Some(overlay) = self.top_trie_changes.get(req.key()) {
-                        self.vm = req.resume_full_value(overlay.as_ref().map(|v| &v[..]));
-                    } else {
-                        self.vm = req.into();
-                        return Verify::StorageGet(StorageGet { inner: self });
-                    }
-                }
-
-                executor::WasmVm::ExternalStorageSet(req) => {
-                    self.top_trie_root_calculation_cache
-                        .as_mut()
-                        .unwrap()
-                        .storage_value_update(req.key(), req.value().is_some());
-                    self.top_trie_changes
-                        .insert(req.key().to_vec(), req.value().map(|v| v.to_vec()));
-                    self.vm = req.resume();
-                }
-
-                executor::WasmVm::ExternalStorageAppend(req) => {
-                    self.top_trie_root_calculation_cache
-                        .as_mut()
-                        .unwrap()
-                        .storage_value_update(req.key(), true);
-
-                    if let Some(current_value) = self.top_trie_changes.get(req.key()) {
-                        let mut current_value = current_value.clone().unwrap_or_default();
-                        append_to_storage_value(&mut current_value, req.value());
-                        self.top_trie_changes
-                            .insert(req.key().to_vec(), Some(current_value));
-                        self.vm = req.resume();
-                    } else {
-                        self.vm = req.into();
-                        return Verify::StorageGet(StorageGet { inner: self });
-                    }
-                }
-
-                executor::WasmVm::ExternalStorageClearPrefix(req) => {
-                    self.vm = req.into();
-                    return Verify::PrefixKeys(PrefixKeys { inner: self });
-                }
-
-                executor::WasmVm::ExternalStorageRoot(req) => {
-                    if self.root_calculation.is_none() {
-                        self.root_calculation = Some(calculate_root::root_merkle_value(Some(
-                            self.top_trie_root_calculation_cache.take().unwrap(),
-                        )));
-                    }
-
-                    match self.root_calculation.take().unwrap() {
-                        calculate_root::RootMerkleValueCalculation::Finished { hash, cache } => {
-                            self.top_trie_root_calculation_cache = Some(cache);
-                            self.vm = req.resume(&hash);
-                        }
-                        calculate_root::RootMerkleValueCalculation::AllKeys(keys) => {
-                            self.vm = req.into();
-                            self.root_calculation =
-                                Some(calculate_root::RootMerkleValueCalculation::AllKeys(keys));
-                            return Verify::PrefixKeys(PrefixKeys { inner: self });
-                        }
-                        calculate_root::RootMerkleValueCalculation::StorageValue(value_request) => {
-                            self.vm = req.into();
-                            // TODO: allocating a Vec, meh
-                            if let Some(overlay) = self
-                                .top_trie_changes
-                                .get(&value_request.key().collect::<Vec<_>>())
-                            {
-                                self.root_calculation =
-                                    Some(value_request.inject(overlay.as_ref()));
-                            } else {
-                                self.root_calculation =
-                                    Some(calculate_root::RootMerkleValueCalculation::StorageValue(
-                                        value_request,
-                                    ));
-                                return Verify::StorageGet(StorageGet { inner: self });
-                            }
-                        }
-                    }
-                }
-
-                executor::WasmVm::ExternalStorageChangesRoot(req) => {
-                    self.vm = req.into();
-                    return Verify::StorageGet(StorageGet { inner: self });
-                }
-
-                executor::WasmVm::ExternalStorageNextKey(req) => {
-                    self.vm = req.into();
-                    return Verify::NextKey(NextKey {
-                        inner: self,
-                        key_overwrite: None,
-                    });
-                }
-
-                executor::WasmVm::ExternalOffchainStorageSet(req) => {
-                    self.offchain_storage_changes
-                        .insert(req.key().to_vec(), req.value().map(|v| v.to_vec()));
-                    self.vm = req.resume();
-                }
-
-                executor::WasmVm::CallRuntimeVersion(req) => {
-                    // The code below compiles the provided WebAssembly runtime code, which is a
-                    // relatively expensive operation (in the order of milliseconds).
-                    // While it could be tempting to use a system cache, this function is expected
-                    // to be called only right before runtime upgrades. Considering that runtime
-                    // upgrades are quite uncommon and that a caching system is rather non-trivial
-                    // to set up, the approach of recompiling every single time is preferred here.
-                    // TODO: number of heap pages?! 1024 is default, but not sure whether that's correct or if we have to take the current heap pages
-                    let vm_prototype = match executor::WasmVmPrototype::new(req.wasm_code(), 1024) {
-                        Ok(w) => w,
-                        Err(_) => {
-                            self.vm = req.resume(Err(()));
-                            continue;
-                        }
-                    };
-
-                    match executor::core_version(vm_prototype) {
-                        Ok((version, _)) => {
-                            // TODO: optimize
-                            self.vm = req.resume(Ok(&parity_scale_codec::Encode::encode(&version)));
-                        }
-                        Err(_) => {
-                            self.vm = req.resume(Err(()));
-                        }
-                    }
-                }
-
-                executor::WasmVm::StartStorageTransaction(tx) => {
-                    self.vm = tx.resume();
-                }
-
-                executor::WasmVm::EndStorageTransaction { resume, rollback } => {
-                    if rollback {
-                        todo!() // TODO:
-                    }
-
-                    self.vm = resume.resume();
-                }
-
-                executor::WasmVm::LogEmit(req) => {
-                    // We add a hardcoded limit to the logs generated by the runtime in order to
-                    // make sure that there is no memory leak. In practice, the runtime should
-                    // rarely log more than a few hundred bytes. This limit is hardcoded rather
-                    // than configurable because it is not expected to be reachable unless
-                    // something is very wrong.
-                    // TODO: optimize somehow? don't create an intermediary String?
-                    let message = req.to_string();
-                    if self.logs.len().saturating_add(message.len()) >= 1024 * 1024 {
-                        return Verify::Finished(Err(Error::LogsTooLong));
-                    }
-
-                    self.logs.push_str(&message);
-                    self.vm = req.resume();
-                }
-            }
-        }
-    }
-}
-
-/// Performs the action described by [`executor::WasmVm::ExternalStorageAppend`] on an encoded
-/// storage value.
-// TODO: remove usage of parity_scale_codec
-fn append_to_storage_value(value: &mut Vec<u8>, to_add: &[u8]) {
-    let curr_len = match <parity_scale_codec::Compact<u64> as parity_scale_codec::Decode>::decode(
-        &mut &value[..],
-    ) {
-        Ok(l) => l,
-        Err(_) => {
-            value.clear();
-            parity_scale_codec::Encode::encode_to(&parity_scale_codec::Compact(1u64), value);
-            value.extend_from_slice(to_add);
-            return;
-        }
-    };
-
-    // Note: we use `checked_add`, as it is possible that the storage entry erroneously starts
-    // with `u64::max_value()`.
-    let new_len = match curr_len.0.checked_add(1) {
-        Some(l) => parity_scale_codec::Compact(l),
-        None => {
-            value.clear();
-            parity_scale_codec::Encode::encode_to(&parity_scale_codec::Compact(1u64), value);
-            value.extend_from_slice(to_add);
-            return;
-        }
-    };
-
-    let curr_len_encoded_size =
-        <parity_scale_codec::Compact<u64> as parity_scale_codec::CompactLen<u64>>::compact_len(
-            &curr_len.0,
-        );
-    let new_len_encoded_size =
-        <parity_scale_codec::Compact<u64> as parity_scale_codec::CompactLen<u64>>::compact_len(
-            &new_len.0,
-        );
-    debug_assert!(
-        new_len_encoded_size == curr_len_encoded_size
-            || new_len_encoded_size == curr_len_encoded_size + 1
-    );
-
-    for _ in 0..(new_len_encoded_size - curr_len_encoded_size) {
-        value.insert(0, 0);
-    }
-
-    parity_scale_codec::Encode::encode_to(&new_len, &mut (&mut value[..new_len_encoded_size]));
-    value.extend_from_slice(to_add);
 }
